Fix unintended buffer cap caused by erroneous u64 to u32 cast
An error in gst_ts_cache_new signature meant that timeshift buffer size
was casted from 64 to 32-bit. This is preventing timeshift buffer going
above 4GB limit and as unintended has now been fixed.

+++ src/tscache.c
@@ -330,11 +330,11 @@ gst_ts_cache_flush (GstTSCache * cache)
  *
  */
 GstTSCache *
-gst_ts_cache_new (gsize size, const gchar * allocator_name)
+gst_ts_cache_new (guint64 size, const gchar * allocator_name)
 {
   GstTSCache *cache;
-  guint nslots;
-  int i;
+  guint64 nslots;
+  guint64 i;
 
   cache = g_new (GstTSCache, 1);
